+++ packages/react/src/planets/Planets.tsx
@@ -5,42 +5,54 @@ import { Vector3 } from "three";
 
 // This should only be accessed through the Planet component, and therefore should always be defined
 // if someone tries to access him outside of a context, it should error somehow
-const PlanetContext = React.createContext<HelloPlanet>({} as HelloPlanet);
+const PlanetContext = React.createContext<HelloPlanet<any, any>>({} as HelloPlanet<any, any>);
 
 export const usePlanet = () => {
   return React.useContext(PlanetContext);
 };
 
-export type PlanetProps<T> = React.PropsWithChildren<{
+export type PlanetProps<T, I> = React.PropsWithChildren<{
   planetProps: HelloPlanetProps;
   numWorkers?: number;
   data?: T;
+  initialData?: I
   lodOrigin: Vector3;
   worker: new () => Worker;
 }>
 
-function PlanetInner<T>(
-  props: PlanetProps<T>,
-  ref: React.ForwardedRef<HelloPlanet<T>>
+function PlanetInner<T, I>(
+  props: PlanetProps<T, I>,
+  ref: React.ForwardedRef<HelloPlanet<T, I>>
 ) {
-  const { children, lodOrigin, worker, data = {}, planetProps, numWorkers } = props
+  const { children, lodOrigin, worker, data, initialData, planetProps, numWorkers } = props
   const planetGroupRef = React.useRef<THREE.Group>(null);
-  const [planetEngine] = React.useState<HelloPlanet>(() => new HelloPlanet<T>(planetProps, worker, numWorkers));
+  const [planetEngine, setPE] = React.useState<HelloPlanet<T, I>>();
 
-
-  React.useImperativeHandle(ref, () => planetEngine as HelloPlanet<T>);
+  React.useImperativeHandle(ref, () => planetEngine as HelloPlanet<T, I>, [planetEngine]);
 
   React.useEffect(() => {
+    if (!planetEngine) {
+      return;
+    }
     planetEngine.planetProps = {
         ...planetEngine.planetProps,
         ...planetProps,
       };
-      // TODO: this needs to be throttled somehow?
-      // otherwise the thread queue gets too long and gets crazy memory issues
-    planetEngine.rebuild(data);
-  }, [planetProps, data]);
+    planetEngine.rebuild(data as T);
+  }, [data, planetEngine]);
+  
+  React.useEffect(() => {
+    const pe = new HelloPlanet<T, I>(planetProps, initialData as I, worker, numWorkers);
+    setPE(pe);
+    return () => {
+      pe.destroy();
+    }
+  }, [planetProps, initialData])
 
-  React.useLayoutEffect(() => {
+  React.useEffect(() => {
+    if (!planetEngine) {
+      return;
+    }
     if (planetGroupRef.current) {
       planetGroupRef.current.add(planetEngine.rootGroup);
     }
@@ -49,23 +61,24 @@ function PlanetInner<T>(
         planetGroupRef.current.remove(planetEngine.rootGroup);
       }
     };
-  }, [planetGroupRef]);
+  }, [planetEngine, planetGroupRef]);
 
   useFrame(() => {
+    if (!planetEngine) {
+      return;
+    }
     if (planetEngine.planetProps) {
-      planetEngine.update(lodOrigin, data);
+      planetEngine.update(lodOrigin, data as T);
     }
   });
 
-  return (
-    <>
-      <PlanetContext.Provider value={planetEngine}>
-        <group ref={planetGroupRef}>{children}</group>
-      </PlanetContext.Provider>
-    </>
-  );
+  return planetEngine ? 
+    <PlanetContext.Provider value={planetEngine as HelloPlanet<T, I>}>
+      <group ref={planetGroupRef}>{children}</group>
+    </PlanetContext.Provider> : null
+  
 }
 
-export const Planet = React.forwardRef(PlanetInner) as <T>(
-  props: PlanetProps<T> & { ref?: React.ForwardedRef<HelloPlanet<T>> }
-) => ReturnType<typeof PlanetInner<T>>;
+export const Planet = React.forwardRef(PlanetInner) as <T, I>(
+  props: PlanetProps<T, I> & { ref?: React.ForwardedRef<HelloPlanet<T, I>> }
+) => ReturnType<typeof PlanetInner<T, I>>;