+++ packages/planets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hello-worlds/planets
 
+## 0.0.8
+
+### Patch Changes
+
+- allow threads to be initialized with initialData
+
 ## 0.0.7
 
 ### Patch Changes